+++ packages/json_serializer/lib/src/util.dart
@@ -1,4 +1,4 @@
-part of belatuk_json_serializer;
+part of '../belatuk_json_serializer.dart';
 
 bool _isPrimitive(value) {
   return value is num || value is bool || value is String || value == null;