+++ packages/pub_sub/example/main.dart
@@ -1,19 +1,20 @@
 import 'dart:io';
 import 'dart:isolate';
-import 'package:belatuk_pub_sub/isolate.dart' as pub_sub;
-import 'package:belatuk_pub_sub/belatuk_pub_sub.dart' as pub_sub;
+import 'package:belatuk_pub_sub/isolate.dart';
+import 'package:belatuk_pub_sub/belatuk_pub_sub.dart';
 
 void main() async {
   // Easily bring up a server.
-  var adapter = pub_sub.IsolateAdapter();
-  var server = pub_sub.Server([adapter]);
+  var adapter = IsolateAdapter();
+  var server = Server([adapter]);
 
   // You then need to create a client that will connect to the adapter.
   // Every untrusted client in your application should be pre-registered.
   //
   // In the case of Isolates, however, those are always implicitly trusted.
+  print("Register Client");
   for (var i = 0; i < Platform.numberOfProcessors - 1; i++) {
-    server.registerClient(pub_sub.ClientInfo('client$i'));
+    server.registerClient(ClientInfo('client$i'));
   }
 
   // Start the server.
@@ -22,6 +23,7 @@ void main() async {
   // Next, let's start isolates that interact with the server.
   //
   // Fortunately, we can send SendPorts over Isolates, so this is no hassle.
+  print("Create Isolate");
   for (var i = 0; i < Platform.numberOfProcessors - 1; i++) {
     await Isolate.spawn(isolateMain, [i, adapter.receivePort.sendPort]);
   }
@@ -32,7 +34,7 @@ void main() async {
 
 void isolateMain(List args) {
   // Isolates are always trusted, so technically we don't need to pass a client iD.
-  var client = pub_sub.IsolateClient('client${args[0]}', args[1] as SendPort);
+  var client = IsolateClient('client${args[0]}', args[1] as SendPort);
 
   // The client will connect automatically. In the meantime, we can start subscribing to events.
   client.subscribe('user::logged_in').then((sub) {