+++ packages/pub_sub/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 6.2.0
+
+* Updated `lints` to 3.0.0
+* Refactored encode/decode message handling into `MessageHandler`
+
+## 6.1.0
+
+* Updated `uuid` to 4.0.0
+
 ## 6.0.0
 
 * Require Dart >= 3.0