+++ packages/range_header/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 6.2.0
+
+* Updated `lints` to 3.0.0
+
+## 6.1.0
+
+* Updated `file` to 7.0.0
+
 ## 6.0.0
 
 * Require Dart >= 3.0